Fix spurious JTAG clock cycles after SPI DMA completes.
Before this change, disabling SPI2 after each DMA exchange while
the JTAG pins were still set as alternate outputs led to the pins
floating, which could lead to TCK becoming set before being driven
low once the JTAG code swaps the pins back to GPIO mode.

Now, the SPI is only disabled after the pins are swapped back to GPIO.

Additionally, we manually set the TDI pin state to the last bit the SPI
controller was transmitting to prevent it reverting to the last state
it was manually set to when we return it to being an output.